foodgram-main/backend/requirements.txt
Django==3.2.3
Pillow==9.0.0
djangorestframework==3.12.4
django-filter==23.1
djoser==2.1.0
environs==11.2.0
gunicorn==20.1.0
jsonschema==4.23.0
psycopg2-binary==2.9.3
pytest==6.2.4
pytest-django==4.4.0
pytest-lazy-fixture==0.6.3
pytest-pythonpath==0.7.3
python-dotenv==1.0.1
